Set up skeleton of main view
The main views consist of a Feed page, where a list of threads will be
displayed; and a Thread page where a single main post and its comments
will be displayed. Redirects from the empty page have also been added,
and models for users, posts, and communities have been written.